requirements.txt
numpy
capytaine
openmeteo_requests
requests_cache
pandas
retry_requests
matplotlib
moviepy
Pillow
